+++ src/ThemeForge.Theme/RegistrySystemThemeProvider.cs
@@ -1,0 +1,67 @@
+// Copyright 2026 Julien Bombled
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using Microsoft.Win32;
+
+namespace ThemeForge.Theme;
+
+/// <summary>Reads the Windows app theme mode from the current user's registry hive.</summary>
+public sealed class RegistrySystemThemeProvider : ISystemThemeProvider, IDisposable
+{
+    private const string PersonalizeKey =
+        @"Software\Microsoft\Windows\CurrentVersion\Themes\Personalize";
+    private const string AppsUseLightThemeValue = "AppsUseLightTheme";
+
+    private bool _subscribed;
+
+    public RegistrySystemThemeProvider()
+    {
+        SystemEvents.UserPreferenceChanged += OnUserPreferenceChanged;
+        _subscribed = true;
+    }
+
+    /// <inheritdoc/>
+    public event EventHandler? Changed;
+
+    /// <inheritdoc/>
+    public SystemThemeMode GetCurrentMode()
+    {
+        using RegistryKey? key = Registry.CurrentUser.OpenSubKey(PersonalizeKey);
+        object? value = key?.GetValue(AppsUseLightThemeValue);
+        if (value is int flag)
+        {
+            return flag == 0 ? SystemThemeMode.Dark : SystemThemeMode.Light;
+        }
+
+        return SystemThemeMode.Unknown;
+    }
+
+    /// <inheritdoc/>
+    public void Dispose()
+    {
+        if (_subscribed)
+        {
+            SystemEvents.UserPreferenceChanged -= OnUserPreferenceChanged;
+            _subscribed = false;
+        }
+    }
+
+    private void OnUserPreferenceChanged(object sender, UserPreferenceChangedEventArgs e)
+    {
+        if (e.Category is UserPreferenceCategory.General or UserPreferenceCategory.Color)
+        {
+            Changed?.Invoke(this, EventArgs.Empty);
+        }
+    }
+}